restructure in prep for async + some async work
