+++ Cargo.toml
@@ -21,6 +21,8 @@ bytes = "1.10.1"
 derive_more = { version = "2.0.1", features = ["display"] }
 serial2 = "0.2.29"
 thiserror = "2.0.12"
+tokio = { version = "1.47.0", features = ["io-util", "net", "time"] }
+tokio-serial = "5.4.5"
 
 [dev-dependencies]
 uuid = { version="1.17.0", features=["v4"] }
